+++ xmake/templates/c++/shared_library_tbox/template.lua
@@ -1,3 +1,6 @@
+-- set name
+set_name("shared_tbox")
+
 -- set description
 set_description("The Shared Library (tbox)")